moved SelfIssuedTokenConstants to SPI
